requirements.txt
torch
Pillow
numpy
huggingface_hub
transformers
accelerate
